+++ SpeedyWeather/src/parameterizations/sea_ice.jl
@@ -2,17 +2,17 @@
 @propagate_inbounds sea_ice_timestep!(vars::Variables, model::PrimitiveEquation) =
     timestep!(vars, model.sea_ice, model)
 
-function variables(::AbstractPrescribedSeaIce)
+function variables(::AbstractSeaIce)
     return (
-        # prescribed sea ice has no step dimension as its not time stepped
+        # prescribed sea ice has no step dimension as its not time stepped, dynamic sea ice defines it with one
         PrognosticVariable(:sea_ice_concentration, GridXY(), namespace = :ocean, desc = "Sea ice concentration", units = "1"),
     )
 end
 
 function variables(::AbstractDynamicSeaIce, model::AbstractModel)
-    nsteps = get_nsteps(model.time_stepping, model)
-    pg = nsteps.prognostic_grid
-    tg = nsteps.tendency_grid
+    nsteps = get_nsteps(model.time_stepping, :ocean)
+    pg = nsteps.prognostic
+    tg = nsteps.tendency
     return (
         PrognosticVariable(:sea_ice_concentration, OceanXYT(pg), namespace = :ocean, desc = "Sea ice concentration", units = "1"),
         TendencyVariable(:sea_ice_concentration, OceanXYT(tg), namespace = :ocean, desc = "Tendency of sea ice concentration", units = "1/s"),
@@ -54,16 +54,16 @@ function timestep!(vars::Variables, sea_ice_model::ThermodynamicSeaIce, model::P
 
     # if ocean does not have an SST tendency use scratch array to write into the void
     dsst = haskey(vars.tendencies.ocean, :sea_surface_temperature) ?
-        get_tendency_step(vars.tendencies.ocean.sea_surface_temperature, model.time_stepping, model.ocean) :
+        get_tendency_step(vars.tendencies.ocean.sea_surface_temperature, time_stepper(model.time_stepping, :ocean), model.ocean) :
         vars.scratch.grid.a_2D
 
     # sea ice concentration written as \aleph yay!
-    dℵ = get_tendency_step(vars.tendencies.ocean.sea_ice_concentration, model.time_stepping, sea_ice_model)
-    sst = get_prognostic_step(vars.prognostic.ocean.sea_surface_temperature, model.time_stepping, model.ocean)
+    dℵ = get_tendency_step(vars.tendencies.ocean.sea_ice_concentration, time_stepper(model.time_stepping, :ocean), sea_ice_model)
+    sst = get_prognostic_step(vars.prognostic.ocean.sea_surface_temperature, time_stepper(model.time_stepping, :ocean), model.ocean)
     
-    @boundscheck size(dsst) == size(dℵ) == size(sst) || throw(BoundsError)
+    @boundscheck size(dsst) == size(dℵ) == size(sst) || throw(BoundsError())
 
-    Δt = time_step(model.time_stepping, vars.prognostic.clock)
+    Δt = time_step(model.time_stepping, :ocean, vars.prognostic.clock)
     (; land_fraction) = model.land_sea_mask
 
     m = sea_ice_model.melt_rate             # melt rate [m²/m²/s/K]